Added Delete Dialog Modal.

+++ EmployeeManagment.Web/Pages/EditEmployeeBase.cs
@@ -7,6 +7,7 @@ using EmployeeManagement.Models;
 using EmployeeManagment.Web.Models;
 using EmployeeManagment.Web.Services;
 using Microsoft.AspNetCore.Components;
+using Venci.Components;
 
 namespace EmployeeManagment.Web.Pages
 {
@@ -29,6 +30,7 @@ namespace EmployeeManagment.Web.Pages
         [Inject] public IMapper Mapper { get; set; }
 
         [Inject] public NavigationManager NavigationManager { get; set; }
+        public ConfirmBase DeleteConfirmation { get; set; }
 
         protected async override Task OnInitializedAsync()
         {
@@ -90,10 +92,17 @@ namespace EmployeeManagment.Web.Pages
             }
         }
 
-        protected async Task Delete_Click()
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            NavigationManager.NavigateTo("/");
+            DeleteConfirmation.Show();
+        }
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                NavigationManager.NavigateTo("/");
+            }
         }
     }
 }
